Smarter auto-indentation after multi-line statements.
This change adds a new indent rule that is smarter when
setting the indent level of a line following a multi-line
statement. Imagine the following code:

  String string1 =
    longString1InitializerOnASeparateLine();
  String string2;

Before this change, jEdit would aling the "string2" line
at the same indent level as the "long..." line before it,
while the correct is to align it with the start of that
statement.

This change requires a new property in the mode file to
tell the rule how to parse statements, otherwise it's
ignored.